This does a single scan over the app definitions.

Benchmarks (both before and after) include a patch pending from @mpark that reduces dependency graph calls.

- All Apps are at the root level
Starting with **_2000_** apps, launching the next 100 sequentially:

**Before**:
_run1_: 2m 07s
_run2_: 2m 23s
**After**:
_run1_: 0m 24s
_run2_: 0m 23s

Starting with **_5000_** apps, launching the next 100 sequentially:
**Before**:
_run1_: 13m 02s
**After**:
_run1_: 0m 50s
_run2_: 0m 54s